- change open_count type to short

+++ wms-opc/src/main/java/cn/tpl/opc/commons/dto/result/CushionInfoDTO.java
@@ -45,7 +45,7 @@ public class CushionInfoDTO extends AbsBaseDTO {
      * 开口数
      */
     @Schema(description = "开口数")
-    private Integer openCount;
+    private Short openCount;
 
     /**
      * 产线
